A lot of changes - read the description
*Moves all currently unused @RePod log generation methods into the
`_core/index/index.php` class

*Thread pruning moved in to deletion class

*Clarified some arbitrary constants for threads per page and pages per
board

*Removed unused `UPDATE_THROTTLING` page method as I can't figure out
how it works. It does a lot of file reading/writing anyway, so it could
probably be done better.
