+++ _core/delete/delete.php
@@ -72,8 +72,8 @@ class SaguaroDelete {
         $result = $mysql->query("SELECT * FROM " . SQLLOG . " WHERE {$conditions} AND board='" . BOARD_DIR . "'");
 
         while ($delrow = $mysql->fetch_assoc($result)) { //This does the resource deletions
-			$this->deleteFile($delrow, $imgonly); //Delete all associated media.
-			$this->deleteResources($delrow, $imgonly); //Delete API, HTML page if necessary
+            $this->deleteFile($delrow, $imgonly); //Delete all associated media.
+            $this->deleteResources($delrow, $imgonly); //Delete API, HTML page if necessary
         }
 
         //Remove posts from the database
@@ -86,20 +86,20 @@ class SaguaroDelete {
     private function deleteFile($post, $imgonly = false) {
         global $mysql;
 
-		$files = json_decode($post['media'], true);
-		$path = realpath("./") . '/' . IMG_DIR;
-		foreach($files as $file) {
-			$delfile = IMG_DIR . $file['localname']; //Path to file
-			$delthumb = THUMB_DIR . $file['localthumbname'];//Path to thumbnail
-			if ($imgonly) {
-				return $mysql->query("UPDATE " . SQLMEDIA . " SET filesize='-1' WHERE no='$no' and board='" . BOARD_DIR ."'");
-			}
-			
-			$mysql->query("DELETE FROM " . SQLMEDIA . " WHERE parent='{$post['no']}' and board='" . BOARD_DIR ."'");
-			
-			if (is_file($delfile)) unlink($delfile); //Delete image
-			if (is_file($delthumb)) unlink($delthumb); //Delete thumbnail
-		}
+        $files = json_decode($post['media'], true);
+        $path = realpath("./") . '/' . IMG_DIR;
+        foreach($files as $file) {
+            $delfile = IMG_DIR . $file['localname']; //Path to file
+            $delthumb = THUMB_DIR . $file['localthumbname'];//Path to thumbnail
+            if ($imgonly) {
+                return $mysql->query("UPDATE " . SQLMEDIA . " SET filesize='-1' WHERE no='$no' and board='" . BOARD_DIR ."'");
+            }
+
+            $mysql->query("DELETE FROM " . SQLMEDIA . " WHERE parent='{$post['no']}' and board='" . BOARD_DIR ."'");
+
+            if (is_file($delfile)) unlink($delfile); //Delete image
+            if (is_file($delthumb)) unlink($delthumb); //Delete thumbnail
+        }
     }
 
     //Delete API files, thread HTML, any other associated files
@@ -121,5 +121,65 @@ class SaguaroDelete {
             @unlink(RES_DIR . $post['no'] . PHP_EXT . '.gz');
         }
         @unlink(RES_DIR . $post['no'] . PHP_EXT);
+    }
+    
+    public function prune_old() { //This does the old thread pruning
+        global $my_log, $mysql;
+        $my_log->update_cache();
+
+        if (PAGES_PER_BOARD >= 1) {
+            $maxposts   = LOG_MAX;
+            $maxthreads = (PAGES_PER_BOARD > 0) ? (PAGES_PER_BOARD * THREADS_PER_PAGE) : 0;
+            //number of pages x how many threads per page
+
+            if ($maxthreads) {
+                $exp_order = (EXPIRE_NEGLECTED == true) ? 'modified' : (defined(EXPIRE_NEGLECTED) ? 'no' : 'modified'); //Legacy config support. For now.
+                
+                $result      = $mysql->query("SELECT no FROM " . SQLLOG . " WHERE sticky=0 AND resto=0 ORDER BY $exp_order ASC");
+                $threadcount = $mysql->num_rows($result);
+                while ($row = $mysql->fetch_array($result) and $threadcount > $maxthreads) {
+                    $this->delete_post($row['no'], 'trim', 0, 1, 1, 0); // imgonly=0, automatic=1, children=1
+                    $threadcount--;
+                }
+                $my_log->update_cache(1); //Force rebuild the cache after batch of deletions is done, instead after every single deletion. 
+                $mysql->free_result($result);
+                // Original max-posts method (note: cleans orphaned posts later than parent posts)
+            } else {
+                // make list of stickies
+                $stickies = array(); // keys are stickied thread numbers
+                $result   = $mysql->query("SELECT no from " . SQLLOG . " where sticky>=1 and resto=0");
+                while ($row = $mysql->fetch_array($result)) {
+                    $stickies[$row['no']] = 1;
+                }
+
+                $result    = $mysql->query("SELECT no,resto,sticky FROM " . SQLLOG . " ORDER BY no ASC");
+                $postcount = $mysql->num_rows($result);
+                while ($row = $mysql->fetch_array($result) and $postcount >= $maxposts) {
+                    // don't delete if this is a sticky thread
+                    if ($row['sticky'] > 0)
+                        continue;
+                    // don't delete if this is a REPLY to a sticky
+                    if ($row['resto'] != 0 && $stickies[$row['resto']] == 1)
+                        continue;
+                    $this->delete_post($row['no'], 'trim', 0, 1, 0, 0); // imgonly=0, automatic=1, children=0
+                    $postcount--;
+                }
+                $mysql->free_result($result);
+            }
+        }
+    }
+
+    public function pruneThread($no) {
+        global $my_log, $mysql;
+        $my_log->update_cache();
+        $maxreplies = EVENT_STICKY_RES;
+
+        $result      = $mysql->query("SELECT no FROM " . SQLLOG . " WHERE resto='$no' ORDER BY time ASC");
+        $repcount = $mysql->num_rows($result);
+        while ($row = $mysql->fetch_array($result) and $repcount >= $maxreplies) {
+            delete_post($row['no'], 'trim', 0, 1, 0, 0); // imgonly=0, automatic=1, children=1
+            $repcount--;
+        }
+        $mysql->free_result($result);
     }
 }
