+++ _core/index/index.php
@@ -152,4 +152,77 @@ class Index {
 
         $this->data = $temp;
     }
+    
+        function generate($type, $no, $inIndex = false) {
+        require_once(CORE_DIR . "/postform.php");
+
+        $dat = '<form name= "delform" action="' . PHP_SELF_ABS . '" method="post">';
+        $foot = '<table align="right"><tr><td class="delsettings" nowrap="nowrap" align="center">
+                <input type="hidden" name="mode" value="usrdel" />' . S_REPDEL . '[<input type="checkbox" name="onlyimgdel" value="on" />' . S_DELPICONLY . ']
+                ' . S_DELKEY . '<input type="password" name="pwd" size="8" maxlength="8" value="" />
+                <input type="submit" value="' . S_DELETE . '" /><input type="button" value="Report" onclick="var o=document.getElementsByTagName(\'INPUT\');for(var i=0;i<o.length;i++)if(o[i].type==\'checkbox\' && o[i].checked && o[i].value==\'delete\') return reppop(\'' . PHP_SELF_ABS . '?mode=report&no=\'+o[i].name+\'\');"></tr></td></form><script>document.delform.pwd.value=l(' . SITE_ROOT . '_pass");</script></td></tr></table>';
+
+        if ($type == "index") {
+            return PostForm::format() . $dat . $this->generate_index($no, $inIndex) . $foot;
+        } elseif ($type == "thread") {
+            return PostForm::format($no) . $dat . $this->generate_thread($no, $inIndex) . $foot;
+        }
+    }
+
+    function generate_index($no, $cacheThreads = false) {
+        $this->update_cache();
+
+        require_once(CORE_DIR . "/index/index.php");
+        $index = new Index;
+        $index_temp = $index->format($no, 0, $cacheThreads);
+
+        $this->thread_cache = $index->thread_cache; //Copy Index thread cache.
+
+        return $index_temp;
+    }
+
+    function generate_thread($no, $inIndex) {
+        $this->update_cache();
+
+        if ($inIndex && $this->thread_cache[$no]) { //Use $this->thread_cache if we want to generate an inIndex thread and it already exists.
+           return $this->thread_cache[$no];
+        } else {
+            require_once(CORE_DIR . "/thread/thread.php"); //Safety.
+            $thread->inIndex = $inIndex;
+            $thread = new Thread;
+
+            return $thread->format($no, true);
+        }
+    }
+
+    function generate_all() {
+        $this->update_cache();
+
+        require_once(CORE_DIR . "/page/page.php");
+        $pageC = new Page;
+
+        $profile = microtime(); //Basic profiling.
+        for ($page = 1; $page <= ceil(count($this->cache['THREADS']) / PAGE_DEF); $page++) {
+            //Generate Index pages.
+            $pageC->headVars['page']['title'] = "/" . BOARD_DIR . "/ - " . TITLE;
+            $temp = $pageC->generate($this->generate("index", $page, false));
+            $logfilename = ($page == 1) ? PHP_SELF2 : ($page - 1). PHP_EXT;
+
+            echo "Writing out Index $page ($logfilename)... ";
+            $this->print_page($logfilename , $temp);
+            echo "Done!<br>";
+        }
+
+        foreach ($this->cache['THREADS'] as $no) {
+                $pageC->headVars['page']['title'] = "/" . BOARD_DIR . "/" . (!empty($this->cache[$no]['sub']) ? " - " . $this->cache[$no]['sub'] : '') . " - " . TITLE;
+                $logfilename = RES_DIR . $no . PHP_EXT;
+                echo "Writing out #$no ($logfilename)... ";
+                $temp = $pageC->generate($this->generate("thread", $no, false));
+
+                $this->print_page($logfilename, $temp);
+                echo "Done!<br>";
+        }
+
+        echo sprintf("<br>Took %f", microtime() - $profile) . " seconds.";
+    }
 }
